# strix-test/Strix-Assets-main/lib/api-zod/src/generated/types/createAccidentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAccidentRequestReportJson } from "./createAccidentRequestReportJson";
import type { ImpactDirection } from "./impactDirection";
import type { ImpactZone } from "./impactZone";
import type { Severity } from "./severity";

export interface CreateAccidentRequest {
  deviceId: string;
  timestamp: Date;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  peakGForce: number;
  impactZone: ImpactZone;
  impactDirection: ImpactDirection;
  speedKmh: number;
  jerkPeak: number;
  approachAngle: number;
  severity: Severity;
  reportJson: CreateAccidentRequestReportJson;
}