# strix-test/Strix-Assets-main/lib/api-zod/src/generated/types/matchAccidentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchAccidentRequest {
  deviceId: string;
  timestamp: Date;
  latitude: number;
  longitude: number;
  approachAngle: number;
}